Move basis algorithms to submodule
