+++ openghg_inversions/basis/algorithms/__init__.py
@@ -1,0 +1,2 @@
+from ._quadtree import get_quadtree_basis as quadtree_algorithm
+from ._weighted import nregion_landsea_basis as weighted_algorithm
